Remove unneeded delimiterSplitter and clean up
